+++ ktorfit-ksp/src/main/kotlin/de/jensklingenberg/ktorfit/reqBuilderExtension/BodyCodeGenerator.kt
@@ -1,7 +1,7 @@
 package de.jensklingenberg.ktorfit.reqBuilderExtension
 
 import de.jensklingenberg.ktorfit.model.ParameterData
-import de.jensklingenberg.ktorfit.model.annotations.Body
+import de.jensklingenberg.ktorfit.model.annotations.ParameterAnnotation.Body
 
 fun getBodyDataText(params: List<ParameterData>): String {
     return params.firstOrNull { it.hasAnnotation<Body>() }?.let {