+++ ktorfit-ksp/src/main/kotlin/de/jensklingenberg/ktorfit/reqBuilderExtension/CustomRequestBuilderCodeGeneration.kt
@@ -1,10 +1,10 @@
 package de.jensklingenberg.ktorfit.reqBuilderExtension
 
 import de.jensklingenberg.ktorfit.model.ParameterData
-import de.jensklingenberg.ktorfit.model.annotations.RequestBuilder
+import de.jensklingenberg.ktorfit.model.annotations.ParameterAnnotation
 
 fun getCustomRequestBuilderText(parameterDataList: List<ParameterData>): String {
-    return (parameterDataList.find { it.hasAnnotation<RequestBuilder>() }?.let {
+    return (parameterDataList.find { it.hasAnnotation<ParameterAnnotation.RequestBuilder>() }?.let {
         it.name + "(this)"
     } ?: "")
 }