+++ ktorfit-ksp/src/test/kotlin/de/jensklingenberg/ktorfit/reqBuilderExtension/ReturnTypeInfoKtTest.kt
@@ -1,6 +1,5 @@
 package de.jensklingenberg.ktorfit.reqBuilderExtension
 
-import de.jensklingenberg.ktorfit.model.ReturnTypeData
 import de.jensklingenberg.ktorfit.requestData.getReturnTypeInfoText
 import org.junit.Assert
 import org.junit.Test